+++ src/helpers/validation/validationConstants.ts
@@ -1,2 +1,2 @@
 export const validModes = new Set(['sequence', 'selector', 'parallel'])
-export const controlConditions = new Set(['and', 'or', 'not'])
+export const controlConditions = new Set(['and', 'or', 'not', 'ensure'])